+++ grpcalchemy/__init__.py
@@ -10,5 +10,4 @@ __email__ = 'guangtian_li@qq.com'
 __version__ = '0.2.0'
 
 from .blueprint import Blueprint, Context
-from .client import Client
 from .server import Server